Local changes to improve performance
There are two parts for this commit:
- The first `Invoke()` call is slow. When the xaml file contains a lot of
VB scripts, JIT compiler will be used, but for **every** compilation, the
compiler starts from the scratch and always tries to resolve missing references
via `MetadataReference.CreateFromFile`, thus it will be very slow.
With some test cache between compilations, one of our test xaml file is
improved from 18s to 3.5s.
- Then I tried calling `Invoke()` 1000 times to see all subsequent workflow
execution behaves. Ideally, they should not degrade from .NET framework.
Profiling shows the WfEventSource is always enabled, so all those tracing
code are run and they takes about 13% of the time.
After force disabling the trace, this 1000 calls are actually faster than
.NET framework.
I am not familiar with the EventSource, but it seems also changed from the
.NET framework implementation, and `EventSource.IsEnabled` always return true,
but it doesn't mean a profiler is attached.