CIVIL-50814: Pre compile all VB expressions first
Now we are reusing the validation services to compile all expressions first.
- A new flag, `IsPreCompilingExpressions` is added.
- Before the real cache data in `WorkflowInstance.ValidateWorkflow()`, we
do this precompiliation, and will cache the `LambdaExpression`s in the
current scope.
- Some unit test failures shows that we cannot cache only by expression text,
so the return type name is also checked. And the activity id cannot be used,
as `DynamicActivity.OnInternalCacheMetadata()` will always create different
`VisualBasicValue` and `VisualBasicReference` activities, thus the cache
will be useless due to different activity id.

Testing shows we can get about 10 seconds improvements for the 21 pkt files.