CIVIL-50814: Reduce no of expressions to compile
If the return type and expression text are the same, we don't need to
queue it for compilation. This is different from validation, where
the activity id is important to trace the errors to activity.
For our precompilation, it's a different situation:
- Most case when execute the workflow, there is no errors.
- It's an optional step, if any error happens, there will be no cached
lambda expressions. And we don't need the error information.